+++ src/main/java/com/jasonvillar/works/register/service/ServiceController.java
@@ -6,11 +6,6 @@ import com.jasonvillar.works.register.service.port.out.ServiceDTO;
 import com.jasonvillar.works.register.service.port.out.ServiceDTOAdapter;
 import com.jasonvillar.works.register.service.port.in.ServiceRequest;
 import com.jasonvillar.works.register.user.User;
-import com.jasonvillar.works.register.user_service.UserService;
-import com.jasonvillar.works.register.user_service.UserServiceService;
-import com.jasonvillar.works.register.user_service.port.in.UserServiceRequest;
-import com.jasonvillar.works.register.user_service.port.out.UserServiceDTO;
-import com.jasonvillar.works.register.user_service.port.out.UserServiceDTOAdapter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -37,10 +32,6 @@ public class ServiceController {
     private final ServiceRequestAdapter serviceRequestAdapter;
 
     private final SecurityUserDetailsService securityUserDetailsService;
-
-    private final UserServiceService userServiceService;
-
-    private final UserServiceDTOAdapter userServiceDTOAdapter;
 
     private final com.jasonvillar.works.register.user.UserService userService;
 
@@ -94,26 +85,6 @@ public class ServiceController {
         return ResponseEntity.ok().body(rowCount);
     }
 
-    @GetMapping(value = "/services/unused/page/{page}/rows/{rows}", produces = "application/json")
-    public ResponseEntity<List<ServiceDTO>> getListUnusedServiceFromUserId(@AuthenticationPrincipal UserDetails userDetails, @PathVariable int page, @PathVariable int rows) {
-        long userId = this.securityUserDetailsService.getAuthenticatedUserId(userDetails);
-        List<ServiceDTO> listDTO = this.service.getUnusedListFromUserId(userId, page - 1, rows).stream().map(serviceDTOAdapter).toList();
-
-        if (listDTO.isEmpty()) {
-            return ResponseEntity.noContent().build();
-        } else {
-            return ResponseEntity.ok().body(listDTO);
-        }
-    }
-
-    @GetMapping(value = "/services/unused/row-count", produces = "application/json")
-    public ResponseEntity<Long> getUnusedRowCountFromUserId(@AuthenticationPrincipal UserDetails userDetails) {
-        long userId = this.securityUserDetailsService.getAuthenticatedUserId(userDetails);
-        long rowCount = this.service.getUnusedRowCountFromUserId(userId);
-
-        return ResponseEntity.ok().body(rowCount);
-    }
-
     @GetMapping(value = "/services/name-like/{name}", produces = "application/json")
     public ResponseEntity<List<ServiceDTO>> getListServiceByNameLike(@AuthenticationPrincipal UserDetails userDetails, @PathVariable String name) {
         long userId = this.securityUserDetailsService.getAuthenticatedUserId(userDetails);
@@ -129,11 +100,12 @@ public class ServiceController {
     @PostMapping(value = "/service")
     public ResponseEntity<Object> saveService(@AuthenticationPrincipal UserDetails userDetails, @Valid @RequestBody ServiceRequest request) {
         long userId = this.securityUserDetailsService.getAuthenticatedUserId(userDetails);
-        Service entity = this.serviceRequestAdapter.toEntity(request);
+        User user = userService.getById(userId);
+        Service entity = this.serviceRequestAdapter.toEntity(request, user);
         String message = this.service.getValidationsMessageWhenCantBeSaved(entity);
 
         if (message.isEmpty()) {
-            entity = this.service.saveWithUser(entity, userId);
+            entity = this.service.save(entity);
             ServiceDTO dto = this.serviceDTOAdapter.apply(entity);
             return new ResponseEntity<>(dto, HttpStatus.CREATED);
         } else {
@@ -141,70 +113,15 @@ public class ServiceController {
         }
     }
 
-    @PostMapping(value = "/service/user")
-    public ResponseEntity<Object> saveUserService(@AuthenticationPrincipal UserDetails userDetails, @Valid @RequestBody UserServiceRequest request) {
-        long userId = this.securityUserDetailsService.getAuthenticatedUserId(userDetails);
-
-        long serviceId = request.serviceId();
-
-        boolean existService = this.service.isExistId(serviceId);
-
-        if (!existService) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Service not founded");
-        }
-
-        UserService entity = UserService.builder()
-                .serviceId(serviceId)
-                .userId(userId)
-                .build();
-
-        boolean existUserService = this.userServiceService.isExistUserIdAndServiceId(userId, serviceId);
-
-        if (existUserService) {
-            return ResponseEntity.status(HttpStatus.FOUND).body("Found existent user service");
-        } else {
-            entity = this.userServiceService.save(entity);
-            entity = this.userServiceService.setServiceEntityIntoUserService(entity);
-            entity = this.userServiceService.setUserEntityIntoUserService(entity);
-
-            UserServiceDTO dto = this.userServiceDTOAdapter.apply(entity);
-            return new ResponseEntity<>(dto, HttpStatus.CREATED);
-        }
-    }
-
     @DeleteMapping(value = "/service/{serviceId}")
     public ResponseEntity<Boolean> deleteServiceToUserId(@AuthenticationPrincipal UserDetails userDetails, @PathVariable long serviceId) {
         long userId = this.securityUserDetailsService.getAuthenticatedUserId(userDetails);
 
-        boolean deleted = this.userServiceService.deleteByServiceIdAndUserId(serviceId, userId);
+        boolean deleted = this.service.deleteByServiceIdAndUserId(serviceId, userId);
         if (deleted) {
             return new ResponseEntity<>(true, HttpStatus.OK);
         } else {
             return new ResponseEntity<>(false, HttpStatus.NOT_FOUND);
         }
-    }
-
-    @PostMapping(value = "/services/delete")
-    public ResponseEntity<List<UserServiceDTO>> deleteServicesBatchToUserId(@AuthenticationPrincipal UserDetails userDetails, @Valid @RequestBody List<Long> serviceIdLongList) {
-        long userId = this.securityUserDetailsService.getAuthenticatedUserId(userDetails);
-
-        List<UserService> userServiceListDeleted = this.userServiceService.deleteByServicesIdAndUserId(serviceIdLongList, userId);
-
-        List<UserServiceDTO> userServiceListDTO = userServiceListDeleted.stream().map(this.userServiceDTOAdapter).toList();
-
-        return new ResponseEntity<>(userServiceListDTO, HttpStatus.OK);
-    }
-
-    @PostMapping(value = "/services/user")
-    public ResponseEntity<List<UserServiceDTO>> addServicesBatchToUserId(@AuthenticationPrincipal UserDetails userDetails, @Valid @RequestBody List<Long> serviceIdLongList) {
-        long userId = this.securityUserDetailsService.getAuthenticatedUserId(userDetails);
-        User user = this.userService.getById(userId);
-
-        List<UserService> userServiceListToSave = this.userServiceService.makeUserServicesFromServiceIdListAndUser(serviceIdLongList, user);
-        List<UserService> userServiceListAdded = this.userServiceService.saveAll(userServiceListToSave);
-
-        List<UserServiceDTO> userServiceListDTO = userServiceListAdded.stream().map(this.userServiceDTOAdapter).toList();
-
-        return new ResponseEntity<>(userServiceListDTO, HttpStatus.OK);
     }
 }