+++ src/main/java/com/jasonvillar/works/register/service/ServiceRepository.java
@@ -1,8 +1,6 @@
 package com.jasonvillar.works.register.service;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import org.springframework.data.domain.Pageable;
@@ -12,32 +10,12 @@ import java.util.Optional;
 @Repository
 public interface ServiceRepository extends JpaRepository<Service, Long> {
     Service findServiceById(long id);
+    List<Service> findAllByUserId(long userId, Pageable pageable);
     Optional<Service> findOptionalById(long id);
-    Optional<Service> findOptionalByName(String name);
-    List<Service> findAllByNameContainingIgnoreCase(String name);
-    Optional<Service> findOptionalByIdAndUserServiceListUserId(long id, long userId);
-    List<Service> findAllByUserServiceListUserId(long userId, Pageable pageable);
-    List<Service> findAllByNameContainingIgnoreCaseAndUserServiceListUserId(String name, long userId);
-    long countByUserServiceListUserId(long userId);
+    Optional<Service> findOptionalByNameAndUserId(String name, long userId);
+    long countByUserId(long userId);
     long count();
-
-    @Query("SELECT s " +
-            "FROM Service s " +
-            "WHERE s.id NOT IN " +
-            "( " +
-            "SELECT us.serviceId " +
-            "FROM UserService us " +
-            "WHERE us.userId = :userId" +
-            ")")
-    List<Service> findAllByUserIdNot(@Param("userId") long userId, Pageable pageable);
-
-    @Query("SELECT COUNT(*) " +
-            "FROM Service s " +
-            "WHERE s.id NOT IN " +
-            "(" +
-            "SELECT us.serviceId " +
-            "FROM UserService us " +
-            "WHERE us.userId = :userId" +
-            ")")
-    long countByUserIdNot(@Param("userId") long userId);
+    Optional<Service> findOptionalByIdAndUserId(long id, long userId);
+    List<Service> findAllByNameContainingIgnoreCaseAndUserId(String name, long userId);
+    boolean deleteByIdAndUserId(long id, long userId);
 }