+++ src/main/java/com/jasonvillar/works/register/service/ServiceService.java
@@ -1,13 +1,12 @@
 package com.jasonvillar.works.register.service;
 
-import com.jasonvillar.works.register.user_service.UserService;
-import com.jasonvillar.works.register.user_service.UserServiceRepository;
+import com.jasonvillar.works.register.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,7 +14,6 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class ServiceService {
     private final ServiceRepository serviceRepository;
-    private final UserServiceRepository userServiceRepository;
 
     public List<Service> getList(int pageNumber, int rows) {
         Pageable page = PageRequest.of(pageNumber, rows, Sort.by("name"));
@@ -24,19 +22,15 @@ public class ServiceService {
 
     public List<Service> getListByUserId(long userId, int pageNumber, int rows) {
         Pageable page = PageRequest.of(pageNumber, rows, Sort.by("name"));
-        return this.serviceRepository.findAllByUserServiceListUserId(userId, page);
+        return this.serviceRepository.findAllByUserId(userId, page);
     }
 
     public long getRowCountByUserId(long userId) {
-        return this.serviceRepository.countByUserServiceListUserId(userId);
+        return this.serviceRepository.countByUserId(userId);
     }
 
     public long getRowCount() {
         return this.serviceRepository.count();
-    }
-
-    public List<Service> getListByNameLike(String name) {
-        return this.serviceRepository.findAllByNameContainingIgnoreCase(name);
     }
 
     public Service getById(long id) {
@@ -47,22 +41,22 @@ public class ServiceService {
         return this.serviceRepository.findOptionalById(id);
     }
 
-    public Optional<Service> getOptionalByName(String name) {
-        return this.serviceRepository.findOptionalByName(name);
+    public Optional<Service> getOptionalByNameAndUserId(String name, long userId) {
+        return this.serviceRepository.findOptionalByNameAndUserId(name, userId);
     }
 
     public boolean isExistId(long id) {
         return this.getOptionalById(id).isPresent();
     }
 
-    public boolean isExistName(String name) {
-        return this.getOptionalByName(name).isPresent();
+    public boolean isExistNameAndUserId(String name, long userId) {
+        return this.getOptionalByNameAndUserId(name, userId).isPresent();
     }
 
     public String getValidationsMessageWhenCantBeSaved(Service entity) {
         StringBuilder message = new StringBuilder();
 
-        if (this.isExistName(entity.getName())) {
+        if (this.isExistNameAndUserId(entity.getName(), entity.getUser().getId())) {
             message.append("name must be unique");
         }
 
@@ -74,40 +68,15 @@ public class ServiceService {
     }
 
     public Optional<Service> getOptionalByIdAndUserId(long id, long userId) {
-        return this.serviceRepository.findOptionalByIdAndUserServiceListUserId(id, userId);
+        return this.serviceRepository.findOptionalByIdAndUserId(id, userId);
     }
 
     public List<Service> getListByNameLikeAndUserId(String name, long userId) {
-        return this.serviceRepository.findAllByNameContainingIgnoreCaseAndUserServiceListUserId(name, userId);
+        return this.serviceRepository.findAllByNameContainingIgnoreCaseAndUserId(name, userId);
     }
 
-    public Service saveWithUser(Service entity, long userId) {
-        Service serviceSaved = this.save(entity);
-
-        UserService userService = UserService.builder()
-                .serviceId(serviceSaved.getId())
-                .userId(userId)
-                .build();
-
-        this.userServiceRepository.save(userService);
-
-        return serviceSaved;
-    }
-
-    public List<Service> getUnusedListFromUserId(long userId, int pageNumber, int rows) {
-        Pageable page = PageRequest.of(pageNumber, rows, Sort.by("name"));
-        return this.serviceRepository.findAllByUserIdNot(userId, page);
-    }
-
-    public long getUnusedRowCountFromUserId(long userId) {
-        return this.serviceRepository.countByUserIdNot(userId);
-    }
-
-    public List<Service> getServiceListInUserService(List<UserService> userService) {
-        List<Service> serviceList = new ArrayList<>();
-        userService.forEach(
-                service -> serviceList.add(service.getService())
-        );
-        return serviceList;
+    @Transactional
+    public boolean deleteByServiceIdAndUserId(long id, long userId) {
+        return this.serviceRepository.deleteByIdAndUserId(id, userId);
     }
 }