+++ src/main/java/com/jasonvillar/works/register/user_service/port/in/UserServiceRequest.java
@@ -1,4 +1,0 @@
-package com.jasonvillar.works.register.user_service.port.in;
-
-public record UserServiceRequest(long serviceId) {
-}