+++ src/test/java/com/jasonvillar/works/register/integration/IntegrationTestsConfig.java
@@ -56,7 +56,6 @@ public class IntegrationTestsConfig {
     @AfterEach
     void afterEach(@Autowired JdbcTemplate jdbcTemplate) {
         JdbcTestUtils.deleteFromTables(jdbcTemplate, "work_register");
-        JdbcTestUtils.deleteFromTables(jdbcTemplate, "user_service");
         JdbcTestUtils.deleteFromTables(jdbcTemplate, "service");
         JdbcTestUtils.deleteFromTables(jdbcTemplate, "client");
     }