+++ src/test/java/com/jasonvillar/works/register/unit/user/UserRepositoryTest.java
@@ -26,7 +26,6 @@ class UserRepositoryTest extends DataJpaTestTemplate {
     @BeforeEach
     void setUp(@Autowired JdbcTemplate jdbcTemplate) {
         JdbcTestUtils.deleteFromTables(jdbcTemplate, "work_register");
-        JdbcTestUtils.deleteFromTables(jdbcTemplate, "user_service");
         JdbcTestUtils.deleteFromTables(jdbcTemplate, "client");
         this.userRepository.findAll().forEach(user -> user.getHasRoleList().forEach(user::removeRole));
         this.userRepository.deleteAll();